- committed cleanups from michael

+++ src/s_stats.c
@@ -19,7 +19,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
  *  USA
  *
- *  $Id: s_stats.c,v 7.22 2002/09/05 01:30:09 db Exp $
+ *  $Id: s_stats.c,v 7.23 2003/01/17 13:00:53 db Exp $
  */
 
 #include "stdinc.h"
@@ -127,8 +127,8 @@ tstats(struct Client *source_p)
 
   sendto_one(source_p, ":%s %d %s :connected %u %u",
              me.name, RPL_STATSDEBUG, source_p->name, 
-	     dlink_list_length(&lclient_list), 
-	     dlink_list_length(&serv_list));
+	     (unsigned int)dlink_list_length(&lclient_list), 
+	     (unsigned int)dlink_list_length(&serv_list));
   sendto_one(source_p, ":%s %d %s :bytes sent %d.%uK %d.%uK",
              me.name, RPL_STATSDEBUG, source_p->name,
              (int)sp->is_cks, sp->is_cbs, (int)sp->is_sks, sp->is_sbs);
